feat:[global] add feacture install with integration in pyposer.json

+++ src/libs/createVenv.py
@@ -9,11 +9,15 @@ def create():
             'click', 'virtualenv', 'six', 'filelock', 'appdirs', 'distlib'
     ]:
         if os.path.exists(f'/usr/lib/python3.9/site-packages/{lib}'):
-
-            shutil.copytree(
-                f'/usr/lib/python3.9/site-packages/{lib}',
-                f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}')
+            if not os.path.exists(
+                    f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}'):
+                shutil.copytree(
+                    f'/usr/lib/python3.9/site-packages/{lib}',
+                    f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}')
         else:
-            shutil.copyfile(
-                f'/usr/lib/python3.9/site-packages/{lib}.py',
-                f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}.py')
+            if not os.path.exists(
+                    f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}.py'
+            ):
+                shutil.copyfile(
+                    f'/usr/lib/python3.9/site-packages/{lib}.py',
+                    f'{os.getcwd()}/venv/lib/python3.9/site-packages/{lib}.py')
